agrego control de excepciones

+++ agregador/src/main/java/ar/edu/utn/frba/dds/services/FuenteService.java
@@ -68,6 +68,9 @@ public class FuenteService {
 
   public ResponseEntity<String> eliminarFuente(String idFuente) {
     Fuente fuente = this.getFuente(idFuente);
+    if (fuente == null) {
+      throw new IllegalArgumentException("Fuente con id " + idFuente + " no existente");
+    }
     fuenteRepository.deleteById(idFuente);
     return ResponseEntity.ok("Operacion de eliminacion completada");
   }
